fix bad imports from src/format/reveal/format-reveal.ts
